requirements.txt
alembic==1.8.0; python_version >= "3.7"
anyio==3.6.1; python_version >= "3.7" and python_full_version >= "3.6.2"
asgiref==3.5.2; python_version >= "3.7"
asyncpg==0.25.0; python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
fastapi==0.78.0; python_full_version >= "3.6.1"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
h11==0.13.0; python_version >= "3.7"
httptools==0.4.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.3; python_version >= "3.7" and python_full_version >= "3.6.2"
mako==1.2.0; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
numpy==1.23.0; python_version >= "3.10" and platform_system == "Linux" and platform_machine == "aarch64" or python_version >= "3.10" or python_version >= "3.9" and platform_system == "Darwin" and platform_machine == "arm64"
opencv-python==4.6.0.66; python_version >= "3.6"
pillow==9.1.1; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pydantic==1.9.1; python_full_version >= "3.6.1"
pydentic==0.0.1.dev3; python_full_version >= "3.6.1"
python-dotenv==0.20.0; python_version >= "3.7"
python-stdnum==1.17; python_full_version >= "3.6.1"
pyyaml==6.0; python_version >= "3.7"
sniffio==1.2.0; python_version >= "3.7" and python_full_version >= "3.6.2"
sqlalchemy==1.4.39; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
toml==0.10.2; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
typing-extensions==4.2.0; python_version >= "3.7" and python_full_version >= "3.6.1"
uvicorn==0.17.6; python_version >= "3.7"
uvloop==0.16.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchgod==0.8.2; python_version >= "3.7"
websockets==10.3; python_version >= "3.7"
